requirements.txt
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
autopep8==2.3.1
beautifulsoup4==4.12.3
boto3==1.35.87
botocore==1.35.90
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cssbeautifier==1.15.1
Django==4.2.17
django-admin-dracula==0.0.3
django-cors-headers==4.6.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-health-check==3.18.3
django-redis==5.4.0
django-silk==5.3.2
django-storages==1.14.4
django-widget-tweaks==1.5.0
djlint==1.36.4
EditorConfig==0.17.0
factory_boy==3.3.1
Faker==33.1.0
frozendict==2.4.6
gprof2dot==2024.6.6
h11==0.14.0
html5lib==1.1
httptools==0.6.4
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
jsbeautifier==1.15.1
json5==0.10.0
lxml==5.3.0
MarkupSafe==3.0.2
msgpack==1.1.0
multitasking==0.0.11
nsepython==2.94
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
patsy==1.0.1
peewee==3.17.8
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
plotly-express==0.4.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
ruff==0.8.4
s3transfer==0.10.4
scipy==1.14.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
statsmodels==0.14.4
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.3
webencodings==0.5.1
websockets==14.1
Werkzeug==3.1.3
yfinance==0.2.51
